Update seeds for recent changes
* Use actual draft data
* Update team names to generic
* Update for new fields
* Standardize with header
* Closes #79
