+++ priv/repo/csv_seed_data/fantasy_leagues.csv
@@ -1,2 +1,3 @@
+Fantasy League Name, Year, Division
 2017 Div A,2017,A
 2017 Div B,2017,B
