+++ priv/repo/csv_seed_data/fantasy_teams.csv
@@ -1,28 +1,29 @@
-Brown,1,14
-Swin/Mara,1,11
-Kintz/Mori,1,9
-Axel/Ken,1,3
-BD,1,8
-Bartch,1,6
-Wilson,1,5
-Miller/Wernz,1,10
-Flo,1,13
-Drew/Gabe,1,7
-Davie/Paul,1,4
-Bro/Heli,1,12
-Jake,1,1
-Zach,1,2
-Vondy/Clark,2,12
-Dan,2,1
-Evan,2,3
-Hughes,2,5
-Erik/Gavin,2,7
-Murphy,2,6
-JakuWOWski,2,4
-Tim/Mike,2,10
-Riss/Ehrman,2,11
-Bailey,2,8
-Nelson,2,9
-Teeter,2,2
-Mick/Jason,2,14
-Tom,2,13
+Team Name,Waiver Position,Dues Paid,Winnings Received,Fantasy League Id
+Cougars,14,0,0,1
+Madd Dogs,10,0,0,1
+Sharks,13,0,0,1
+Jay Hawks,6,0,0,1
+Wild Cats,6,0,0,2
+Black Mambas,1,0,0,1
+Bengal Tigers,2,0,0,1
+Huskies,12,0,0,1
+Rosie's Riveters,3,0,0,1
+Junior Mints,10,0,0,2
+She-Unit,5,0,0,1
+Black Widows,9,0,0,1
+Blitz Babes,7,0,0,2
+Bare Exposure,9,0,0,2
+Pink Cougars,7,0,0,1
+End Zone Divas,11,0,0,1
+Girl Power,11,0,0,2
+Victorious Secret,4,0,0,1
+Air Farce,1,0,0,2
+Team Jagermeister,13,0,0,2
+Hoosier Daddys,8,0,0,2
+Ocho Stinko,5,0,0,2
+No Punt Intended,8,0,0,1
+Loose Ends,2,0,0,2
+Little Jerry Seinfelds,4,0,0,2
+Deep Threat,3,0,0,2
+Back that Pass Up,12,0,0,2
+Fumble This,14,0,0,2
