+++ priv/repo/csv_seed_data/owners.csv
@@ -1,0 +1,3 @@
+Fantasy Team, User
+4,1
+2,2
