+++ priv/repo/csv_seed_data/waivers.csv
@@ -1,0 +1,4 @@
+Fantasy Team Id,Add Fantasy Player Id,Drop Fantasy Player Id,Status,Process At
+7,306,294,pending,2016-10-13 10:00:00
+12,,268,pending,2016-10-10 10:05:00
+16,395,,pending,2016-10-13 10:06:00
